[SPARK-42163][SQL] Fix schema pruning for non-foldable array index or map key
### What changes were proposed in this pull request?

In parquet schema pruning, we use SelectedField to try to extract the field that is used in a struct. It looks through GetArrayItem/GetMapItem, but when doing so, it ignores the index/key, which may itself be a struct field. If it is a struct field that is not selected in some other expression, and another field of the same attribute is selected, then pruning will drop the field, resulting in an optimizer error.

This change modifies SelectedField to only look through GetArrayItem/GetMapItem if the index/key argument is foldable. The equivalent code for `ElementAt` was already doing the same thing, so this just makes them consistent.

In principle, we could continue to traverse through these expressions, we'd just need to make sure that the index/key expression was also surfaced to column pruning as an expression that needs to be examined. But this seems like a fairly non-trivial change to the design of the SelectedField class.

There is some risk that the current approach could result in a regression e.g. if there is an existing GetArrayItem that is being successfully pruned, where a non-foldable index argument happens to not trigger an error (because it is not a struct field, or it is preserved due to some other expression).

### Why are the changes needed?

Allows queries that previously would fail in the optimizer to pass.

### Does this PR introduce _any_ user-facing change?

Yes, as described above, there could be a performance regression if a query was previously pruning through a GetArrayItem/GetMapItem, and happened to not fail.

### How was this patch tested?

Unit test included in patch, fails without the patch and passes with it.

Closes #39718 from cashmand/fix_selected_field.

Authored-by: cashmand <david.cashman@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit 16cfa092d2c23a1cca45acd62e2095a72cb27d86)
Signed-off-by: Wenchen Fan <wenchen@databricks.com>