[SPARK-42344][K8S] Change the default size of the CONFIG_MAP_MAXSIZE
The default size of the CONFIG_MAP_MAXSIZE should not be greater than 1048576

### What changes were proposed in this pull request?
This PR changed the default size of the  CONFIG_MAP_MAXSIZE from 1572864（1.5 MiB) to 1048576（1.0 MiB）

### Why are the changes needed?
When a job is submitted by the spark to the K8S with a configmap , The Spark-Submit will call the K8S‘s POST API "api/v1/namespaces/default/configmaps". And the size of the configmaps will be validated by this K8S API,the max value shoud not be greater than 1048576.
In the previous comment，the explain in https://etcd.io/docs/v3.4/dev-guide/limit/ is:
"etcd is designed to handle small key value pairs typical for metadata. Larger requests will work, but may increase the latency of other requests. By default, the maximum size of any request is 1.5 MiB. This limit is configurable through --max-request-bytes flag for etcd server."
This explanation is from the perspective of etcd ，not K8S.
So I think the default value of the configmap in Spark should not be greate than 1048576.

### Does this PR introduce _any_ user-facing change?
Yes.
Generally, the size of the configmap will not exceed 1572864 or even 1048576.
So the problem solved here may not be perceived by users.

### How was this patch tested?
local test

Closes #39884 from ninebigbig/master.

Authored-by: Yan Wei <ninebigbig@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 9ac46408ec943d5121bbc14f2ce0d8b2ff453de5)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>