[SPARK-42369][CORE] Fix constructor for java.nio.DirectByteBuffer
### What changes were proposed in this pull request?

In the latest JDK, the constructor `DirectByteBuffer(long, int)` was replaced with `DirectByteBuffer(long, long)`. We just want to support both by probing for the legacy one first and falling back to the newer one second.

This change is completely transparent for the end-user, and makes sure Spark works transparently on the latest JDK as well.

### Why are the changes needed?

It's required to support the latest versions of the JDK.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

It was built and ran locally against the newest JDK.

Closes #39909 from luhenry/patch-2.

Authored-by: Ludovic Henry <git@ludovic.dev>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>