[SPARK-42656][CONNECT] Adding SCALA REPL shell script for JVM client
### What changes were proposed in this pull request?
Adding a simple script to start the Scala client in the Scala REPL. As well as a script to start the spark connect server for the client to connect to.

### Why are the changes needed?
Make the JVM client more easy to be used.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Manually tested.

Closes #40257 from zhenlineo/jshell.

Authored-by: Zhen Li <zhenlineo@users.noreply.github.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>
(cherry picked from commit 7505e5cdd12661dd5c96bfd004185bf6fc4eb33b)
Signed-off-by: Herman van Hovell <herman@databricks.com>