[SPARK-42496][CONNECT][DOCS] Adding Spark Connect to the Spark 3.4 documentation
### What changes were proposed in this pull request?
Adding a Spark Connect overview page to the Spark 3.4 documentation and a short section on the Spark overview page with a link to it.

### Why are the changes needed?
The first version of Spark Connect is released as part of Spark 3.4.0 and this adds an overview for it to the documentation.

### Does this PR introduce _any_ user-facing change?
Yes, the user facing documentation is updated.

### How was this patch tested?
Built the doc website locally and tested the pages.
SKIP_SCALADOC=1 SKIP_RDOC=1 bundle exec jekyll build

index.html
![index html](https://user-images.githubusercontent.com/112507318/224800134-7d0093bb-261a-4abe-a3da-b1e8784854c1.png)

spark-connect-overview.html
![spark-connect-overview html](https://user-images.githubusercontent.com/112507318/224800167-8f6308e0-2e0e-4b15-81a8-5bd9da12d0ce.png)

Closes #40324 from allanf-db/spark_connect_docs.

Authored-by: Allan Folting <allan.folting@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit c7b9b42efdd577cb7ea41752fb6e73444462d0ce)
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>