[MINOR][DOCS] Remove SparkSession constructor invocation in the example
### What changes were proposed in this pull request?

This PR proposes to Remove SparkSession constructor invocation in the example.

While I am here, I piggyback and add an example of Spark Connect.

### Why are the changes needed?

SparkSession's constructor is not meant to be exposed to the end users. This is also hidden in Scala side.

### Does this PR introduce _any_ user-facing change?

Yes, it removes the usage of SparkSession constructor in the user facing docs.

### How was this patch tested?

Linters should verify the changes in the CI.

Closes #40505 from HyukjinKwon/minor-docs-session.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit 54afa2bfffb46d80cbe9f815a37c3d3325648ef3)
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>