[SPARK-42903][PYTHON][DOCS] Avoid documenting None as as a return value in docstring
### What changes were proposed in this pull request?

This PR proposes to remove None as as a return value in docstring.

### Why are the changes needed?

To be consistent with the current documentation. Also, it's idiomatic to don't specify the return for `return None`.

### Does this PR introduce _any_ user-facing change?

Yes, it changes the user-facing documentation.

### How was this patch tested?

Doc build in the CI should verify them.

Closes #40530 from HyukjinKwon/SPARK-42903.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit 07301b4ab96365cfee1c6b7725026ef3f68e7ca1)
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>