[SPARK-44000][SQL] Add hint to disable broadcasting and replicating one side of join
### What changes were proposed in this pull request?

This PR adds a new internal join hint to disable broadcasting and replicating one side of join.

### Why are the changes needed?

These changes are needed to disable broadcasting and replicating one side of join when it is not permitted, such as the cardinality check in MERGE operations in PR #41448.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This PR comes with tests. More tests are in #41448.

Closes #41499 from aokolnychyi/spark-44000.

Authored-by: aokolnychyi <aokolnychyi@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit d88633ada5eb73e8876acaa2c2a53b9596f2acdd)