[SPARK-43975][SQL] DataSource V2: Handle UPDATE commands for group-based sources
### What changes were proposed in this pull request?

This PR adds support for group-based data sources in `RewriteUpdateTable` and builds on top of the existing logic.

The UPDATE command for group-based is executed differently based on whether there is a subquery in the condition. If there is no subquery, there is simply a projection to update matching rows. If there is a subquery, the rule would assign a union of two scans: one to find matching rows and one to carry over unmatched records in matched groups. The union is required so that subqueries can be rewritten as joins via `RewritePredicateSubquery`. In the future, we may provide a more optimized execution if the subquery is small. That said, the current approach would scale better and should be the default.

### Why are the changes needed?

These changes are needed per SPIP SPARK-35801.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This PR comes with tests. There are more tests in `AlignUpdateAssignmentsSuite`, which was merged earlier.

Closes #41600 from aokolnychyi/spark-43975.

Authored-by: aokolnychyi <aokolnychyi@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 209722817e27d03b9d55ba5d1ebff1f99c531ce7)