[SPARK-42454][SQL] SPJ: encapsulate all SPJ related parameters in BatchScanExec
Pull out the SPJ-related attribute of BatchScanExec into a case class

We plan to have further evolution of SPJ parameters to support more SPJ features.  So we want to stabilize the definition of BatchScanExec to not have to touch the many places in the code that it is pattern-matched/unapplied, etc..

No

Existing unit test to verify no behavior change.

Closes #41990 from szehon-ho/spj_refactor.

Authored-by: Szehon Ho <szehon.apache@gmail.com>
Signed-off-by: Chao Sun <sunchao@apple.com>