rdar://113352087 Fix artifact filter for publishing images (#1837)
We need to pull the artifact URL in order to publish an image based specifically on the published distribution.  The same build is shared across both snapshot and release image builds (i.e. YAML `release_image_build`).

This could be done at the end of the upstream job, but the distribution is architecture independent and a multi-arch image build would end up building multiple distributions.