rdar://114213400 Bump spark-call-home-listener to 0.2.31 (#1855)
### What changes were proposed in this pull request?
Bump the `spark-call-home-listener` to 0.2.31.

### Why are the changes needed?
This version addresses API differences that are specific to Spark 3.4.0.  This differences caused failures.

During testing, it was also identified that there were problems introduced due to the circular dependencies between Spark and `spark-call-home`.  The builds were updated to exclude `spark-call-home-listener` from Spark.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
The gradle build was updated to add new Spark 3.4 profiles, and the build-specific profiles were used in all building.  This demonstrated the API issue, causing the build to fail.  In addition new PR builds were introduced to perform this testing as part of each PR.  In addition, the tests were supplemented by builds and tests of Spark with the PR-built `spark-call-home-listener` were added to ensure that Spark was unaffected.

In addition, the latest Spark release image was patched to replace the `spark-call-home-listener` jar, and this was used to run `spark-perf` benchmarks.

### Was this patch authored or co-authored using generative AI tooling?
No

### Patch
https://github.pie.apple.com/pie-spark/spark-call-home/compare/f2eb2d0b2b...a93f3d08bb

- feat: Apply spark and scala profile (#115)
- fix: Handle default Spark for macros and exclude circular dependency (#116)
- build: Replace gala-platform-runtime with ubi9-minimal/java11-runtime (#118)
- build: Update base image for docker deploy (#119)