[SPARK-45592][SPARK-45282][SQL][3.4] Correctness issue in AQE with InMemoryTableScanExec
### What changes were proposed in this pull request?

This backports https://github.com/apache/spark/pull/43435 SPARK-45592  to the 3.4 branch. This is because it was already reported there as SPARK-45282 but it required enabling some extra configuration to hit the bug.

### Why are the changes needed?

Fix correctness issue.

### Does this PR introduce _any_ user-facing change?

Yes, fixing correctness issue.

### How was this patch tested?

New tests based on the reproduction example in SPARK-45282

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #43729 from eejbyfeldt/SPARK-45282.

Authored-by: Emil Ejbyfeldt <eejbyfeldt@liveintent.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>