[SPARK-46353][CORE] Refactor to improve `RegisterWorker` unit test coverage
This PR aims to improve the unit test coverage for `RegisterWorker` message handling.

- Add `handleRegisterWorker` helper method which is testable easily.
- Add new unit tests for three conditional branches.

It's easily to test and improve. We can add more tests in this way in the future.

No. This is a refactoring on the main code and only additions to the test methods.

Pass the CIs.

No.

Closes #44284 from dongjoon-hyun/SPARK-46353.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>