[SPARK-46786][K8S] Fix `MountVolumesFeatureStep` to use `ReadWriteOncePod` instead of `ReadWriteOnce`
This PR aims to fix a duplicated volume mounting bug by using `ReadWriteOncePod` instead of `ReadWriteOnce`.

This bug fix is based on the stable K8s feature which is available since v1.22.

- [KEP-2485: ReadWriteOncePod PersistentVolume AccessMode](https://github.com/kubernetes/enhancements/blob/master/keps/sig-storage/2485-read-write-once-pod-pv-access-mode/README.md)
- https://kubernetes.io/docs/concepts/storage/persistent-volumes/#access-modes
    - v1.22 Alpha
    - v1.27 Beta
    - v1.29 Stable

For the record, the minimum K8s version of GKE/EKS/AKE is **v1.24** as of today and the latest v1.29 is supported like the following.
- [2024.01 (GKE Regular Channel)](https://cloud.google.com/kubernetes-engine/docs/release-schedule)
- [2024.02 (AKE GA)](https://learn.microsoft.com/en-us/azure/aks/supported-kubernetes-versions?tabs=azure-cli#aks-kubernetes-release-calendar)

This is a bug fix.

Pass the CIs with the existing PV-related tests.

No.

Closes #44817 from dongjoon-hyun/SPARK-46786.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>
(cherry picked from commit 45ec74415a4a89851968941b80c490e37ee88daf)
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>