[SPARK-46881][CORE] Support `spark.deploy.workerSelectionPolicy`
### What changes were proposed in this pull request?

This PR aims to introduce a new configuration, `spark.deploy.workerSelectionPolicy`, in order to allow users to decide a
policy to assign executors on one of the assignable workers.
- `CORES_FREE_ASC`: Choose a worker with the least free cores
- `CORES_FREE_DESC`: Choose a worker with the most free cores (**default**)
- `MEMORY_FREE_ASC`: Choose a worker with the least free memory
- `MEMORY_FREE_DESC`: Choose a worker with the most free memory
- `WORKER_ID`: Choose a worker with the smallest worker id

### Why are the changes needed?

This is very helpful for users to manage Spark Standalone cluster dynamically based on their workload and underlying cluster management policy.

### Does this PR introduce _any_ user-facing change?

No. The default policy, `CORES_FREE_DESC`, follows the existing behavior.

### How was this patch tested?

Pass the CIs with the newly added test cases.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #44906 from dongjoon-hyun/SPARK-46881.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>