[SPARK-47185][SS][TESTS] Increase timeout between actions in KafkaContinuousSourceSuite
### What changes were proposed in this pull request?

This PR proposes to increase the timeout between between actions in `KafkaContinuousSourceSuite`.

### Why are the changes needed?

In Mac OS build, those tests fail indeterministically, see
- https://github.com/apache/spark/actions/runs/8054862135/job/22000404856
- https://github.com/apache/spark/actions/runs/8040413156/job/21958488693
- https://github.com/apache/spark/actions/runs/8032862212/job/21942732320
- https://github.com/apache/spark/actions/runs/8024427919/job/21937366481

`KafkaContinuousSourceSuite` is specifically slow in Mac OS. Kafka producers send the messages correctly, but the consumers can't get the messages for some reasons. You can't get the offsets for long time. This is not an issue in micro batch but I fail to identify the difference.

I just decided to increase the timeout between actions for now. This is more just a workaround.

### Does this PR introduce _any_ user-facing change?

No, test-only.

### How was this patch tested?

Manually tested in my Mac.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45283 from HyukjinKwon/SPARK-47185.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>